+++ src/worker/routes/index.ts
@@ -5,6 +5,7 @@ export { handleIndexDocs } from "./index-docs";
 export { handleRetrieve } from "./retrieve";
 export { handleStatus } from "./status";
 export { handleDeleteDocset, handleGetDocset } from "./docset";
+export { handleListPages } from "./pages";
 export { handleRefresh, handleRefreshAll } from "./refresh";
 export { 
   handleRegisterSession,